Change JobInterviewStageId type from int to short
The JobInterviewStageId property in JobInterviewStageModel was updated to use the short data type instead of int. No other modifications were made.

Change JobInterviewStageId and JobStatusId to short

Changed the data types of JobInterviewStageId in JobInterviewStageModel and JobStatusId in JobStatusModel from int to short for consistency and potential memory optimization.

+++ JobUpdatesAPI/Models/JobInterviewStageModel.cs
@@ -2,7 +2,7 @@
 
 public class JobInterviewStageModel
 {
-    public int JobInterviewStageId { get; set; }
+    public short JobInterviewStageId { get; set; }
     public int CurrentStage {  get; set; } = 0; 
     public int InterviewStages { get; set; }
